fixed the error where a live bet would concat like a string, updated schema to include numeric values for decimals

+++ database/schema.sql
@@ -17,7 +17,7 @@ CREATE TABLE "public"."users" (
 CREATE TABLE "public"."bets" (
 	"betId" serial NOT NULL,
 	"gameId" TEXT NOT NULL,
-	"betAmount" integer NOT NULL,
+	"betAmount" NUMERIC(12, 2) NOT NULL,
 	"betType" TEXT NOT NULL,
 	"createdAt" timestamptz NOT NULL default now(),
   "gameStart" timestamptz NOT NULL,
@@ -36,7 +36,7 @@ CREATE TABLE "public"."spreadBets" (
 	"homeTeam" TEXT NOT NULL,
 	"awayTeam" TEXT NOT NULL,
 	"price" integer NOT NULL,
-	"points" integer NOT NULL,
+	"points" NUMERIC(12, 1) NOT NULL,
 	"userId" integer NOT NULL,
 	"potentialWinnings" float4 NOT NULL,
 	CONSTRAINT "spreadBets_pk" PRIMARY KEY ("spreadBetId")
@@ -65,7 +65,7 @@ CREATE TABLE "public"."totals" (
 	"homeTeam" TEXT NOT NULL,
 	"awayTeam" TEXT NOT NULL,
 	"price" integer NOT NULL,
-	"points" integer NOT NULL,
+	"points" NUMERIC(12, 1) NOT NULL,
 	"userId" integer NOT NULL,
 	"potentialWinnings" float4 NOT NULL,
 	CONSTRAINT "totals_pk" PRIMARY KEY ("totalsId")
